Provide fallbacks for paths of environment variables using a singleton.
The former implementation of ghi_1368 didn't work upstream because of the following messages:

> QCoreApplication::applicationDirPath: Please instantiate the QApplication object first

That makes sense, GlobalAttributes was not bound to an instance of QApplication but I tried to access that pretty early. Not sure why things seemed to have worked for me, though. This commit changes the former implementation to an approach using a singleton, so that one instance of GlobalAttributes can be created at app-start and simply used pretty similar like in the past everywhere else.

THINGS DON'T WORK CURRENTLY!

> Exception code=0xc00000fd flags=0x0 at 0x000000006E5C3BC4

Not sure how to debug this, so am providing this PoC upstream first to discuss the general approach. This fixes #1368.